{{cookiecutter.django_project_name}}/requirements.txt
Django==3.0.4
psycopg2-binary==2.8.4
sentry-sdk==0.14.2
dj-database-url==0.5.0
boto==2.49.0
boto3==1.12.13
django-storages==1.9.1
djangorestframework==3.11.0
django-debug-toolbar==2.2
django-webpack-loader==0.7.0
